Relocate PlayerIdentity and synchronize access
