+++ core/src/main/java/dev/pgm/community/Community.java
@@ -1,6 +1,6 @@
 package dev.pgm.community;
 
-import static dev.pgm.community.nick.PlayerIdentity.PLAYER_IDENTITY;
+import static dev.pgm.community.nick.identity.PlayerIdentity.PLAYER_IDENTITY;
 
 import dev.pgm.community.commands.graph.CommunityCommandGraph;
 import dev.pgm.community.events.CommunityEvent;
